08_python/requirements.txt
pandas
numpy
matplotlib
pyarrow
google-cloud-bigquery
pandas-gbq
jupyter
